+++ moshmosh/__init__.py
@@ -1,6 +1,7 @@
 from .ast_compat import ast
 from .extension_register import *
 from .extensions import template_python
+from .extensions import lazy_import
 from .extensions import pattern_matching
 from .extensions import scoped_operators
 from .extensions import pipelines