+++ moshmosh/extensions/lazy_import/__init__.py
@@ -1,0 +1,2 @@
+from .runtime import *
+from .main import *