+++ moshmosh/extensions/pattern_matching/main.py
@@ -143,9 +143,7 @@ class GenMatch(ast.NodeTransformer):
         suite.reverse()
         return suite
 
-
 class PatternMatching(Extension):
-    __slots__ = ('tokens', )
     identifier = 'pattern-matching'
 
     def pre_rewrite_src(self, io: StringIO):
@@ -159,3 +157,11 @@ class PatternMatching(Extension):
 
     def __init__(self, token='match'):
         self.token = token
+
+    def prior(self, other):
+        # must be prior to quick-lambda,
+        # otherwise, the wildcard might not work well.
+
+        from moshmosh.extensions.quick_lambdas import QuickLambda
+        it = isinstance(other, QuickLambda) and other.token == "_"
+        return it